+++ src/main/java/com/bst/red_green_blue/pojo/WechatNews.java
@@ -9,11 +9,11 @@ public class WechatNews {
 
     private String description;
 
-    private Integer time;
+    private Long time;
 
     private String teamId;
 
-    public WechatNews(String id, String title, String picture, String description, Integer time, String teamId) {
+    public WechatNews(String id, String title, String picture, String description, Long time, String teamId) {
         this.id = id;
         this.title = title;
         this.picture = picture;
@@ -58,11 +58,11 @@ public class WechatNews {
         this.description = description == null ? null : description.trim();
     }
 
-    public Integer getTime() {
+    public Long getTime() {
         return time;
     }
 
-    public void setTime(Integer time) {
+    public void setTime(Long time) {
         this.time = time;
     }
